refactor(schema): Separate derivation from evolution checks
- Move accepted-set compatibility into a dedicated schema evolution module.
- Reuse the shared inclusion engine for derivation admission while preserving derivation-specific rules.
- Update casting, store, trait validation, and tests to use the focused APIs.
